+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphtrace/config
 
+## 1.6.2
+
+### Patch Changes
+
+- @graphtrace/shared@1.6.2
+
 ## 1.6.1
 
 ### Patch Changes