+++ packages/indexer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphtrace/indexer
 
+## 1.6.2
+
+### Patch Changes
+
+- @graphtrace/config@1.6.2
+- @graphtrace/shared@1.6.2
+- @graphtrace/storage@1.6.2
+
 ## 1.6.1
 
 ### Patch Changes